BE: [Feat] Question user 인증 권한 부여 구현 complete

+++ server/src/main/java/pre_002/stackOverFlow_Clone/question/controller/QuestionController.java
@@ -19,10 +19,13 @@ import pre_002.stackOverFlow_Clone.question.entity.Question;
 import pre_002.stackOverFlow_Clone.question.mapper.QuestionMapper;
 import pre_002.stackOverFlow_Clone.question.service.QuestionService;
 import pre_002.stackOverFlow_Clone.user.mapper.UserMapper;
+import pre_002.stackOverFlow_Clone.user.service.UserService;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
+import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @RequestMapping("/")
@@ -33,6 +36,7 @@ public class QuestionController {
     private final QuestionMapper questionMapper;
     private final AnswerService answerService;
     private final AnswerMapper answerMapper;
+    private final UserService userService;
     private final UserMapper userMapper;
 
     // 전체 질문 조회
@@ -62,26 +66,35 @@ public class QuestionController {
     }
 
     // 질문 등록
-    @PostMapping("/questionlist")
-    public Long postQuestion(@Valid @RequestBody QuestionDto.Post post) {
+    @PostMapping("/auth/questionlist")
+    public Long postQuestion(@Valid @RequestBody QuestionDto.Post post,
+                             Principal principal) {
 
-        Question question = questionService.postQuestion(questionMapper.postToQuestion(post));
+//        User user = userService.findVerifiedUserByEmail(principal.getName());
+
+        Question question = questionService.postQuestion(questionMapper.postToQuestion(post), principal);
 
 //        System.out.println(question.getQuestionContents());
 //        DetailQuestionResponseDto response = questionMapper.questionToResponse(question);
 
 //        return new ResponseEntity<>(new SingleResponseDto<>(response), HttpStatus.CREATED);
+
         return question.getQuestionId();
     }
 
     // 질문 수정
-    @PatchMapping("/questionlist/{question-id}")
+    @PatchMapping("/auth/questionlist/{question-id}")
     public Long patchQuestion(@PathVariable("question-id") @Positive Long questionId,
-                                        @Valid @RequestBody QuestionDto.Patch patch) {
+                                        @Valid @RequestBody QuestionDto.Patch patch,
+                                        Principal principal) {
 
         patch.setQuestionId(questionId);
-
         Question question = questionService.patchQuestion(questionMapper.patchToQuestion(patch));
+
+        if (!Objects.equals(principal.getName(), question.getUser().getEmail())) {
+
+            throw new BusinessLogicException(ExceptionCode.FORBIDDEN_USER);
+        }
 
 //        QuestionDto.Response response = questionMapper.questionToResponse(question);
 //
@@ -90,9 +103,17 @@ public class QuestionController {
         return question.getQuestionId();
     }
 
-    @DeleteMapping("/questionlist/{question-id}")
-    public void deleteQuestion(@PathVariable("question-id") @Positive Long questionId) {
-        
+    @DeleteMapping("/auth/questionlist/{question-id}")
+    public void deleteQuestion(@PathVariable("question-id") @Positive Long questionId,
+                               Principal principal) {
+
+        Question question = questionService.getQuestion(questionId);
+
+        if (!Objects.equals(principal.getName(), question.getUser().getEmail())) {
+
+            throw new BusinessLogicException(ExceptionCode.FORBIDDEN_USER);
+        }
+
         questionService.delete(questionId);
 
 //        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
